Revert "Revise MoonPhase calculation for more correct brightness"
This reverts commit baf28173c42b1de539ac6a586a22c4ea7484fd17.

+++ src/logic/MoonPhase.hpp
@@ -8,6 +8,6 @@ class MoonPhase
     enum { Full, Waning, Waxing, New };
 
 public:
-    static double fromDateTime(const QDateTime &t);
-    static QString formatToString(double phase);
+    static int fromDateTime(const QDateTime &t);
+    static QString formatToString(int phase);
 };
